requirements.txt
alpaca-py>=0.21.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
pytz>=2023.3
schedule>=1.2.0
pytest>=7.0.0
requests>=2.28.0
# ib_insync>=0.9.86  # Uncomment when switching to IB
